Correct possible memory leak
In the parseFile method, on line 80 a new char array is created for moData.
So if an error is found the memory must be freed before exiting.
Instead of deleting the memory, the moData_ variable could have been set to moData and let clearData() do is job.
I prefer this solution since it's easier to understand.
This error was reported by Cppcheck.